requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
configparser==5.2.0
cssselect2==0.4.1
defusedxml==0.7.1
idna==3.3
Pillow==9.0.0
pycparser==2.21
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tinycss2==1.1.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uuid==1.30
webencodings==0.5.1
